Refactor models and utils
